Diana-AI-backend/requirements.txt
# Core API
fastapi>=0.115
uvicorn[standard]>=0.30
pydantic>=2.7
pydantic-settings>=2.3
python-dotenv>=1.0
httpx>=0.27
orjson>=3.10

# Web utils
python-multipart>=0.0.9
aiofiles>=23.2

# Embeddings / NLP
sentence-transformers>=3.0
torch>=2.3 ; platform_system != "Darwin"  # CPU está bien en Windows
numpy>=1.26
tqdm>=4.66

# Vector store
faiss-cpu>=1.8.0

# Opcionales
scikit-learn>=1.4
rapidfuzz>=3.9
loguru>=0.7
